pyproject.toml
[project]
name = "pqcow"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "msgspec==0.18.6",
    "pqcow-liboqs==0.10.1.dev0",
    "cryptography==43.0.3",
    "websockets==14.1",
]

[tool.uv]
package = true

[project.optional-dependencies]
server = [
    "aiosqlite==0.20.0",
]
dev = [
    "isort==5.13.2",
    "pre-commit==4.0.1",
    "ruff==0.8.0",
    "bandit==1.7.10",
]
lint = [
    "mypy==1.13.0",
    "types-pytz==2024.2.0.20241003",
]
uvloop = [
    "uvloop==0.21.0; sys_platform == 'linux' or sys_platform == 'darwin'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pqcow"]
artifacts = [
    "/oqs",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "/tests",
    "/liboqs-python",
]

[tool.isort]
py_version = 312
src_paths = ["pqcow", "tests"]
line_length = 100
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = true
split_on_trailing_comma = false
single_line_exclusions = ["."]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_gitignore = true
extend_skip = ["__pycache__"]
extend_skip_glob = []

[tool.ruff]
src = ["pqcow", "tests"]
line-length = 100
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "__pypackages__",
    "__pycache__",
    "*.pyi",
    "venv",
    ".venv",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "A003",
    "ANN002", "ANN003", "ANN401",
    "C901",
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D203", "D205", "D212",
    "ERA001",
    "FA100", "FA102",
    "FBT001", "FBT002",
    "FIX002",
    "I001",
    "PLR0911", "PLR0912", "PLR0913", "PLR0915", "PLR5501",
    "PLW0120",
    "RUF001",
    "TD002", "TD003"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
mypy_path = "."
packages = ["pqcow"]
plugins = []
allow_redefinition = true
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
extra_checks = true
follow_imports_for_stubs = true
ignore_missing_imports = false
namespace_packages = true
no_implicit_optional = true
no_implicit_reexport = true
pretty = true
show_absolute_path = true
show_error_codes = true
show_error_context = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

disable_error_code = [
    "no-redef",
]

exclude = [
    "\\.?venv",
    "\\.idea",
    "\\.tests?",
]
